Add 2019 agenda
Create the CSVs using the agenda from the conference spreadsheet:
https://docs.google.com/spreadsheets/d/18Y8rsF2Am3JPypLbkoXquCMzkRky58cEPQo9nLz3xdE/edit#gid=8

Put the agenda table generation code in a shortcode snippet, so we can
use it twice on the same page. We probably could (should?) do this
using a partial with arguments... but this works.

The archive pages can now use the 'archive' shortcode (which is just an
iframe), or the 'agenda' shortcode to generate the table from CSVs.

Add the footer to the archive pages. Would be nice if this could
include the actual committee members from each year.