Replace iframe agendas for 2014-2018
The script imports the HTML from the old website, parses it, and
creates a CSV of the agenda for each day. That CSV is used by the
'agendatable' shortcode to create the page in the same way as the
2019-present agendas.

The script also updates the presentation URLs to point to their S3
counterparts, when possible.